+++ pyccel/parser/syntactic.py
@@ -323,7 +323,10 @@ class SyntaxParser(BasicParser):
         if txt.startswith("$"):
             env = txt[1:].strip()
             if env.startswith("omp"):
-                if env.endswith("&") or env.endswith(r"\"):
+                if env.endswith("&") or env.endswith("\\"):
+                    if env.endswith("&"):
+                        errors.report("Using & as an OpenMP continuation character is deprecated and will be removed in v2.5.",
+                                      severity="warning", symbol=stmt)
                     if self._multiline_comment_in_progress:
                         self._multiline_comment_in_progress.append(env[3:-1])
                     else: